+++ src/io/index/filter.rs
@@ -20,7 +20,7 @@ pub fn should_keep(read: &SequenceRecord, opts: &FilterOpts) -> bool {
         Some(v) => opts.quality.contains(v),
         None => true, // if there is no quality, we do NOT want to remove
     };
-    let len_good = opts.len.contains(read.num_bases() as f64);
+    let len_good = opts.len.contains(read.num_bases() as f32);
 
     quality_good && len_good
 }
